Hide the server commonad on not supported platform
