+++ cmd/server/root_windows.go
@@ -1,0 +1,8 @@
+//go:build windows
+// +build windows
+
+package server
+
+func init() {
+	isNotSupported = true
+}
